requirements.txt
selenium
PyGithub
python-dotenv